bighead-trading/requirements.txt
# Avantis Trading Bot Dependencies

aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# For live trading (install separately when ready)
# avantis_trader_sdk
# eth-account
